Infrastructure/ai-model/requirements.txt
kafka-python
influxdb-client[ciso]

darts
pandas
scikit-learn
numpy
xgboost
lightgbm